+++ data/settings-datastore/src/main/java/com/google/jetpackcamera/data/settingsdatastore/SettingsModule.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2023 The Android Open Source Project
+ * Copyright (C) 2026 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.jetpackcamera.settings
+package com.google.jetpackcamera.data.settingsdatastore
 
+import com.google.jetpackcamera.settings.SettingsRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
 
 /**
  * Dagger [Module] for settings data layer.
@@ -32,21 +32,4 @@ interface SettingsModule {
     fun bindsSettingsRepository(
         localSettingsRepository: LocalSettingsRepository
     ): SettingsRepository
-
-    @Binds
-    @Singleton
-    fun bindsSettableConstraintsRepository(
-        settableConstraintsRepository: SettableConstraintsRepositoryImpl
-    ): SettableConstraintsRepository
-
-    /**
-     * ConstraintsRepository without setter.
-     *
-     * This is the same instance as the singleton SettableConstraintsRepository, but does not
-     * have the ability to update the constraints.
-     */
-    @Binds
-    fun bindsConstraintsRepository(
-        constraintsRepository: SettableConstraintsRepository
-    ): ConstraintsRepository
 }
