+++ data/settings-datastore/src/main/proto/com/google/jetpackcamera/settings/jca_settings.proto
