+++ settings.gradle.kts
@@ -56,3 +56,4 @@ include(":core:model")
 include(":ui:uistate:postcapture")
 include(":ui:uistateadapter:postcapture")
 include(":core:camera:postprocess")
+include(":data:settings-datastore")
